Create Rails 6.1 pool proxy and split files
Default ActiveRecord DB config now returns an object instead of a hash
so we have to use a new way of getting the hash config.
